@@ -2,14 +2,14 @@
 
   <PropertyGroup>
     <TargetFramework>netstandard2.0</TargetFramework>
-	<LangVersion>12.0</LangVersion>
+	<LangVersion>13.0</LangVersion>
 	<Nullable>enable</Nullable>
 	<EnforceExtendedAnalyzerRules>true</EnforceExtendedAnalyzerRules>
 	<AnalyzerLanguage>cs</AnalyzerLanguage>
 	<IsRoslynComponent>true</IsRoslynComponent>
 	<AllowUnsafeBlocks>true</AllowUnsafeBlocks>
 	<Title>QPlayer.SourceGenerator</Title>
-	<Version>0.1.34-pre</Version>
+	<Version>0.1.36-pre</Version>
 	<Authors>Thomas Mathieson</Authors>
 	<Copyright>Copyright © Thomas Mathieson 2026</Copyright>
 	<PackageProjectUrl>https://github.com/space928/QPlayer</PackageProjectUrl>
@@ -20,6 +20,10 @@
 	<PackageLicenseExpression>MIT</PackageLicenseExpression>
 	<PackageRequireLicenseAcceptance>True</PackageRequireLicenseAcceptance>
   </PropertyGroup>
+
+  <ItemGroup>
+    <Compile Include="..\QPlayer\Utilities\TemporaryList.cs" Link="TemporaryList.cs" />
+  </ItemGroup>
 
   <ItemGroup>
     <None Include=".\README.md">
